refactor: FE directory
