--- frontend/src/components/AdminPage/MainSelect.js
@@ -1,1 +1,0 @@
-// 관리자페이지 메인 메뉴 선정하는 부분
