+++ frontend/src/components/Review/ReviewReport.js
@@ -1,6 +1,6 @@
 import React from "react";
-import * as config from "../config";
 import styled from "@emotion/styled";
+import * as config from "../../config";
 
 const ReportButton = styled.span`
 	width: 25px;
