+++ frontend/src/pages/RestaurantDetailPage/MenuSelector.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
-import "rsuite/dist/rsuite-no-reset.min.css";
 import styled from "@emotion/styled";
+import "rsuite/dist/rsuite-no-reset.min.css";
 import { Cascader } from "rsuite";
 
 const MenuSelectorContainer = styled.div`
