+++ frontend/src/pages/RestaurantDetailPage/ModalMenuTable.js
@@ -1,5 +1,5 @@
-import styled from "@emotion/styled";
 import React, { useEffect, useState } from "react";
+import styled from "@emotion/styled";
 import * as config from "../../config";
 
 const Table = styled.table`
